Created app and organized fizzbuzz files
