--- fizzbuzz.py
@@ -1,10 +1,0 @@
-def fizzbuzz(num):
-    if (num % 3 == 0) and (num % 5 == 0):
-        response = 'fizzbuzz'
-    elif num % 3 == 0:
-        response = 'fizz'
-    elif num % 5 == 0:
-        response = 'buzz'
-    else:
-        response = num
-    return response
